marketing-analytics/activation/gmp-googleads-connector/requirements.txt

google-api-python-client
googleads==19.0.1
google-cloud-firestore
zeep>=2.5.0
requests>=2.0.0,<3.0.0
xmltodict>=0.9.2,<1.0.0
google-auth>=1.0.0,<2.0.0
google-auth-oauthlib>=0.0.1,<1.0.0
